Move test fixtures to conftest.py
